+++ packages/jazz-tools/src/tools/coValues/coFeed.ts
@@ -14,10 +14,8 @@ import type {
   AnonymousJazzAgent,
   CoValue,
   CoValueClass,
-  CoreAccountSchema,
   Group,
   ID,
-  InstanceOfSchema,
   RefsToResolve,
   RefsToResolveStrict,
   Resolved,
@@ -31,10 +29,8 @@ import {
   CoValueBase,
   ItemsSym,
   Ref,
-  RegisteredSchemas,
   SchemaInit,
   accessChildById,
-  anySchemaToCoSchema,
   coField,
   ensureCoValueLoaded,
   inspect,
